requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pandas
openpyxl
PyPDF2
pdfplumber
jsondiff
requests
python-dotenv
openai
xlrd
werkzeug
slowapi
